update sse implementation
