+++ dgemm-naive.c
@@ -13,23 +13,25 @@ LDLIBS = -lrt -Wl,--start-group $(MKLROOT)/lib/intel64/libmkl_intel_lp64.a $(MKL
 
 */
 
-const char* dgemm_desc = "Naive, three-loop dgemm.";
+const char *dgemm_desc = "Naive, three-loop dgemm.";
 
 /* This routine performs a dgemm operation
  *  C := C + A * B
  * where A, B, and C are lda-by-lda matrices stored in column-major format.
- * On exit, A and B maintain their input values. */    
-void square_dgemm (int n, double* A, double* B, double* C)
+ * On exit, A and B maintain their input values. */
+void square_dgemm(int n, double *A, double *B, double *C)
 {
-  /* For each row i of A */
-  for (int i = 0; i < n; ++i)
-    /* For each column j of B */
-    for (int j = 0; j < n; ++j) 
-    {
-      /* Compute C(i,j) */
-      double cij = C[i+j*n];
-      for( int k = 0; k < n; k++ )
-	cij += A[i+k*n] * B[k+j*n];
-      C[i+j*n] = cij;
-    }
+    /* For each row i of A */
+    for (int i = 0; i < n; ++i)
+        /* For each column j of B */
+        for (int j = 0; j < n; ++j)
+        {
+            /* Compute C(i,j) */
+            double cij = C[i + j * n];
+            for (int k = 0; k < n; k++)
+            {
+                cij += A[i + k * n] * B[k + j * n];
+            }
+            C[i + j * n] = cij;
+        }
 }
